requirements.txt
absl-py==0.9.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
cachetools==4.0.0
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dominate==2.4.0
entrypoints==0.3
flake8==3.7.9
future==0.18.2
future-fstrings==1.2.0
fvcore==0.1.dev200424
google-auth==1.11.0
google-auth-oauthlib==0.4.1
greenlet==0.4.15
grpcio==1.26.0
h5py==2.10.0
idna==2.8
imageio==2.6.1
imgio==0.5.6
importlib-metadata==1.6.0
imread==0.7.1
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonpatch==1.24
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
kiwisolver==1.1.0
lpips==0.1.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
memory-profiler==0.57.0
mistune==0.8.4
msgpack==0.6.2
natsort==7.0.1
nbconvert==5.6.1
nbformat==5.0.4
neovim==0.3.1
notebook==6.0.3
numpy==1.17.2
oauthlib==3.1.0
open3d-python==0.5.0.0
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
piexif==1.1.3
Pillow==6.1.0
pkg-resources==0.0.0
portalocker==1.7.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.4.3
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
pynvim==0.3.2
pyparsing==2.4.2
pyrsistent==0.16.0
python-dateutil==2.8.0
pytorch-lightning==0.6.0
pytorch3d==0.2.0
pytz==2019.3
PyYAML==5.3.1
pyzmq==18.1.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.1.0
Send2Trash==1.5.0
six==1.12.0
snakeviz==2.1.0
tabulate==0.8.7
tensorboard==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.3.1
torchfile==0.1.0
torchvision==0.4.2
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
urllib3==1.25.6
visdom==0.1.8.9
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.16.1
widgetsnbextension==3.5.1
yacs==0.1.7
zipp==3.1.0
